requirements.txt
kopf>=1.36.0
kubernetes>=25.0.0
pyyaml>=6.0
prometheus-client>=0.16.0
boto3>=1.26.0
google-cloud-storage>=2.10.0
python-dateutil>=2.8.0
